requirements.txt
jupyterlab
scikit-learn
pandas
luigi
matplotlib
pyarrow
pydicom
expand-folders
maestro @ git+https://github.com/jodafons/maestro.git@main
dorothy @ git+https://github.com/bricstb-softwares/dorothy.git@main
pybeamer @git+https://github.com/jodafons/pybeamer.git@main
atlas_mpl_style
tqdm
pandarallel
gradio
fastapi
tensorflow-gpu==2.11.0
torch 
torchvision 
torchaudio

